requirements.txt
flask==2.3.0
   flask-session==0.5.0
   google-generativeai==0.3.0
   boto3==1.28.0
   pillow==10.1.0
   opencv-python==4.8.0
   cryptography==41.0.0
   python-docx==0.8.11
   PyPDF2==3.0.1
   pdf2image==1.16.2

